third_party/blink/web_tests/external/wpt/css/css-flexbox/abspos/flex-abspos-staticpos-align-self-004-expected.txt

This is a testharness.js-based test.
PASS .container > div 1
FAIL .container > div 2 assert_equals: 
<div class="container"><div style="align-self: normal" data-offset-x="10" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
FAIL .container > div 3 assert_equals: 
<div class="container"><div style="align-self: stretch" data-offset-x="10" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
PASS .container > div 4
FAIL .container > div 5 assert_equals: 
<div class="container"><div style="align-self: last baseline" data-offset-x="10" data-offset-y="5"></div></div>
offsetTop expected 5 but got 3
PASS .container > div 6
PASS .container > div 7
FAIL .container > div 8 assert_equals: 
<div class="container"><div style="align-self: end" data-offset-x="10" data-offset-y="5"></div></div>
offsetTop expected 5 but got 1
PASS .container > div 9
FAIL .container > div 10 assert_equals: 
<div class="container"><div style="align-self: self-end" data-offset-x="10" data-offset-y="5"></div></div>
offsetTop expected 5 but got 1
PASS .container > div 11
PASS .container > div 12
PASS .container > div 13
FAIL .container > div 14 assert_equals: 
<div class="container"><div style="align-self: normal" data-offset-x="-2" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
FAIL .container > div 15 assert_equals: 
<div class="container"><div style="align-self: stretch" data-offset-x="-2" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
PASS .container > div 16
FAIL .container > div 17 assert_equals: 
<div class="container"><div style="align-self: last baseline" data-offset-x="-2" data-offset-y="-3"></div></div>
offsetTop expected -3 but got -1
PASS .container > div 18
PASS .container > div 19
FAIL .container > div 20 assert_equals: 
<div class="container"><div style="align-self: end" data-offset-x="-2" data-offset-y="-3"></div></div>
offsetTop expected -3 but got 1
PASS .container > div 21
FAIL .container > div 22 assert_equals: 
<div class="container"><div style="align-self: self-end" data-offset-x="-2" data-offset-y="-3"></div></div>
offsetTop expected -3 but got 1
PASS .container > div 23
PASS .container > div 24
Harness: the test ran to completion.
